simple inheritance is more straightforward
